+++ docs/Changelog.md
@@ -297,6 +297,7 @@ Editor features:
   * Swap Hyper Inferno and Seeker Swarm spellbooks back to their 1999 positions.
   * Potions can now be picked up even when your health/mana is full.
   * Disable interface indicators when out of mana and when a spellbook is tier 2.
+  * Fix Mama Spider and Mush The Shroom to not drop cash due to being considered bosses.
 
 ### v4.2 (2025-05-25)